Refactor fraud detection integration: Moved Gradio app import to a new module and removed obsolete deployment script and Gradio app files.
