+++ src/students/joshsalako/api/main.py
@@ -5,10 +5,10 @@ import gradio as gr
 from agents.chatbot.llm_gradio import llm_chat
 from api.models import UpdateUserRequest, UserRequest
 from api.safe_eval import safe_eval
-from dagster.ml_fraud.gradio_app import iface as fraud_app
 from fastapi import FastAPI, HTTPException
 from fastapi.openapi.docs import get_swagger_ui_html
 from gradioapp.app import app as demo
+from gradioapp.fraud_app import iface as fraud_app
 from gradioapp.heart_disease_app import heart_app
 from gradioapp.image_assignment import image_app
